Single-page blog setup using CSS grid and flexbox
This change sets up a single page blog using CSS grid and flexbox functionality. The page consists of 3 regions - header, section, and footer. We apply grid functionality to the outer layout to render the three regions. We apply flexbox to the contents of the section region that contains the blog post contents.

We also introduce media queries to set different style rules for different screen sizes. For screens larger than 900px, we increase the font size from the default size medium to large.